% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{AntennaCharacteristics}
\alias{AntennaCharacteristics}
\title{Google Spectrum Database API Objects 
API for spectrum-management functions.}
\usage{
AntennaCharacteristics(height = NULL, heightType = NULL,
  heightUncertainty = NULL)
}
\arguments{
\item{height}{The antenna height in meters}

\item{heightType}{If the height is required, then the height type (AGL for above ground level or AMSL for above mean sea level) is also required}

\item{heightUncertainty}{The height uncertainty in meters}
}
\value{
AntennaCharacteristics object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:47:29
filename: /Users/mark/dev/R/autoGoogleAPI/googlespectrumv1explorer.auto/R/spectrum_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
AntennaCharacteristics Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Antenna characteristics provide additional information, such as the antenna height, antenna type, etc. Whether antenna characteristics must be provided in a request depends on the device type and regulatory domain.
}

