% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{DatabaseSpec}
\alias{DatabaseSpec}
\title{DatabaseSpec Object}
\usage{
DatabaseSpec(name = NULL, uri = NULL)
}
\arguments{
\item{name}{The display name for a database}

\item{uri}{The corresponding URI of the database}
}
\value{
DatabaseSpec object
}
\description{
DatabaseSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This message contains the name and URI of a database.
}

