% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{DbUpdateSpec}
\alias{DbUpdateSpec}
\title{DbUpdateSpec Object}
\usage{
DbUpdateSpec(databases = NULL)
}
\arguments{
\item{databases}{A required list of one or more databases}
}
\value{
DbUpdateSpec object
}
\description{
DbUpdateSpec Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This message is provided by the database to notify devices of an upcoming change to the database URI.
}
