% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{DeviceCapabilities}
\alias{DeviceCapabilities}
\title{DeviceCapabilities Object}
\usage{
DeviceCapabilities(frequencyRanges = NULL)
}
\arguments{
\item{frequencyRanges}{An optional list of frequency ranges supported by the device}
}
\value{
DeviceCapabilities object
}
\description{
DeviceCapabilities Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Device capabilities provide additional information that may be used by a device to provide additional information to the database that may help it to determine available spectrum. If the database does not support device capabilities it will ignore the parameter altogether.
}
