% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{DeviceDescriptor}
\alias{DeviceDescriptor}
\title{DeviceDescriptor Object}
\usage{
DeviceDescriptor(etsiEnDeviceCategory = NULL,
  etsiEnDeviceEmissionsClass = NULL, etsiEnDeviceType = NULL,
  etsiEnTechnologyId = NULL, fccId = NULL, fccTvbdDeviceType = NULL,
  manufacturerId = NULL, modelId = NULL, rulesetIds = NULL,
  serialNumber = NULL)
}
\arguments{
\item{etsiEnDeviceCategory}{Specifies the ETSI white space device category}

\item{etsiEnDeviceEmissionsClass}{Specifies the ETSI white space device emissions class}

\item{etsiEnDeviceType}{Specifies the ETSI white space device type}

\item{etsiEnTechnologyId}{Specifies the ETSI white space device technology identifier}

\item{fccId}{Specifies the device's FCC certification identifier}

\item{fccTvbdDeviceType}{Specifies the TV Band White Space device type, as defined by the FCC}

\item{manufacturerId}{The manufacturer's ID may be required by the regulatory domain}

\item{modelId}{The device's model ID may be required by the regulatory domain}

\item{rulesetIds}{The list of identifiers for rulesets supported by the device}

\item{serialNumber}{The manufacturer's device serial number; required by the applicable regulatory domain}
}
\value{
DeviceDescriptor object
}
\description{
DeviceDescriptor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The device descriptor contains parameters that identify the specific device, such as its manufacturer serial number, regulatory-specific identifier (e.g., FCC ID), and any other device characteristics required by regulatory domains.
}

