% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{DeviceOwner}
\alias{DeviceOwner}
\title{DeviceOwner Object}
\usage{
DeviceOwner(operator = NULL, owner = NULL)
}
\arguments{
\item{operator}{The vCard contact information for the device operator is optional, but may be required by specific regulatory domains}

\item{owner}{The vCard contact information for the individual or business that owns the device is required}
}
\value{
DeviceOwner object
}
\description{
DeviceOwner Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This parameter contains device-owner information required as part of device registration. The regulatory domains may require additional parameters.All contact information must be expressed using the structure defined by the vCard format specification. Only the contact fields of vCard are supported:  - fn: Full name of an individual - org: Name of the organization - adr: Address fields - tel: Telephone numbers - email: Email addresses  Note that the vCard specification defines maximum lengths for each field.
}

