% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{DeviceValidity}
\alias{DeviceValidity}
\title{DeviceValidity Object}
\usage{
DeviceValidity(deviceDesc = NULL, isValid = NULL, reason = NULL)
}
\arguments{
\item{deviceDesc}{The descriptor of the device for which the validity check was requested}

\item{isValid}{The validity status: true if the device is valid for operation, false otherwise}

\item{reason}{If the device identifier is not valid, the database may include a reason}
}
\value{
DeviceValidity object
}
\description{
DeviceValidity Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The device validity element describes whether a particular device is valid to operate in the regulatory domain.
}

