% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{EventTime}
\alias{EventTime}
\title{EventTime Object}
\usage{
EventTime(startTime = NULL, stopTime = NULL)
}
\arguments{
\item{startTime}{The inclusive start of the event}

\item{stopTime}{The exclusive end of the event}
}
\value{
EventTime object
}
\description{
EventTime Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The start and stop times of an event. This is used to indicate the time period for which a spectrum profile is valid.Both times are expressed using the format, YYYY-MM-DDThh:mm:ssZ, as defined in RFC3339. The times must be expressed using UTC.
}

