% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{FrequencyRange}
\alias{FrequencyRange}
\title{FrequencyRange Object}
\usage{
FrequencyRange(channelId = NULL, maxPowerDBm = NULL, startHz = NULL,
  stopHz = NULL)
}
\arguments{
\item{channelId}{The database may include a channel identifier, when applicable}

\item{maxPowerDBm}{The maximum total power level (EIRP)—computed over the corresponding operating bandwidth—that is permitted within the frequency range}

\item{startHz}{The required inclusive start of the frequency range (in Hertz)}

\item{stopHz}{The required exclusive end of the frequency range (in Hertz)}
}
\value{
FrequencyRange object
}
\description{
FrequencyRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A specific range of frequencies together with the associated maximum power level and channel identifier.
}

