% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{GeoLocation}
\alias{GeoLocation}
\title{GeoLocation Object}
\usage{
GeoLocation(confidence = NULL, point = NULL, region = NULL)
}
\arguments{
\item{confidence}{The location confidence level, as an integer percentage, may be required, depending on the regulatory domain}

\item{point}{If present, indicates that the geolocation represents a point}

\item{region}{If present, indicates that the geolocation represents a region}
}
\value{
GeoLocation object
}
\description{
GeoLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This parameter is used to specify the geolocation of the device.
}
