% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{GeoLocationEllipse}
\alias{GeoLocationEllipse}
\title{GeoLocationEllipse Object}
\usage{
GeoLocationEllipse(center = NULL, orientation = NULL,
  semiMajorAxis = NULL, semiMinorAxis = NULL)
}
\arguments{
\item{center}{A required geo-spatial point representing the center of the ellipse}

\item{orientation}{A floating-point number that expresses the orientation of the ellipse, representing the rotation, in degrees, of the semi-major axis from North towards the East}

\item{semiMajorAxis}{A floating-point number that expresses the location uncertainty along the major axis of the ellipse}

\item{semiMinorAxis}{A floating-point number that expresses the location uncertainty along the minor axis of the ellipse}
}
\value{
GeoLocationEllipse object
}
\description{
GeoLocationEllipse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A 'point' with uncertainty is represented using the Ellipse shape.
}
