% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{GeoLocationPoint}
\alias{GeoLocationPoint}
\title{GeoLocationPoint Object}
\usage{
GeoLocationPoint(latitude = NULL, longitude = NULL)
}
\arguments{
\item{latitude}{A required floating-point number that expresses the latitude in degrees using the WGS84 datum}

\item{longitude}{A required floating-point number that expresses the longitude in degrees using the WGS84 datum}
}
\value{
GeoLocationPoint object
}
\description{
GeoLocationPoint Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single geolocation on the globe.
}

