% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{GeoLocationPolygon}
\alias{GeoLocationPolygon}
\title{GeoLocationPolygon Object}
\usage{
GeoLocationPolygon(exterior = NULL)
}
\arguments{
\item{exterior}{When the geolocation describes a region, the exterior field refers to a list of latitude/longitude points that represent the vertices of a polygon}
}
\value{
GeoLocationPolygon object
}
\description{
GeoLocationPolygon Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A region is represented using the polygonal shape.
}
