% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{GeoSpectrumSchedule}
\alias{GeoSpectrumSchedule}
\title{GeoSpectrumSchedule Object}
\usage{
GeoSpectrumSchedule(location = NULL, spectrumSchedules = NULL)
}
\arguments{
\item{location}{The geolocation identifies the location at which the spectrum schedule applies}

\item{spectrumSchedules}{A list of available spectrum profiles and associated times}
}
\value{
GeoSpectrumSchedule object
}
\description{
GeoSpectrumSchedule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The schedule of spectrum profiles available at a particular geolocation.
}
