% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.R
\name{paws.getSpectrum}
\alias{paws.getSpectrum}
\title{Requests information about the available spectrum for a device at a location. Requests from a fixed-mode device must include owner information so the device can be registered with the database.}
\usage{
paws.getSpectrum(PawsGetSpectrumRequest)
}
\arguments{
\item{PawsGetSpectrumRequest}{The \link{PawsGetSpectrumRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{http://developers.google.com/spectrum}{Google Documentation}

Other PawsGetSpectrumRequest functions: \code{\link{PawsGetSpectrumRequest}}
}

