% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.R
\name{paws.getSpectrumBatch}
\alias{paws.getSpectrumBatch}
\title{The Google Spectrum Database does not support batch requests, so this method always yields an UNIMPLEMENTED error.}
\usage{
paws.getSpectrumBatch(PawsGetSpectrumBatchRequest)
}
\arguments{
\item{PawsGetSpectrumBatchRequest}{The \link{PawsGetSpectrumBatchRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{http://developers.google.com/spectrum}{Google Documentation}

Other PawsGetSpectrumBatchRequest functions: \code{\link{PawsGetSpectrumBatchRequest}}
}
