% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.R
\name{paws.notifySpectrumUse}
\alias{paws.notifySpectrumUse}
\title{Notifies the database that the device has selected certain frequency ranges for transmission. Only to be invoked when required by the regulator. The Google Spectrum Database does not operate in domains that require notification, so this always yields an UNIMPLEMENTED error.}
\usage{
paws.notifySpectrumUse(PawsNotifySpectrumUseRequest)
}
\arguments{
\item{PawsNotifySpectrumUseRequest}{The \link{PawsNotifySpectrumUseRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{http://developers.google.com/spectrum}{Google Documentation}

Other PawsNotifySpectrumUseRequest functions: \code{\link{PawsNotifySpectrumUseRequest}}
}

