% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.R
\name{paws.register}
\alias{paws.register}
\title{The Google Spectrum Database implements registration in the getSpectrum method. As such this always returns an UNIMPLEMENTED error.}
\usage{
paws.register(PawsRegisterRequest)
}
\arguments{
\item{PawsRegisterRequest}{The \link{PawsRegisterRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item 
}

Set \code{options(googleAuthR.scopes.selected = c()}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{http://developers.google.com/spectrum}{Google Documentation}

Other PawsRegisterRequest functions: \code{\link{PawsRegisterRequest}}
}

