% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{PawsGetSpectrumBatchRequest}
\alias{PawsGetSpectrumBatchRequest}
\title{PawsGetSpectrumBatchRequest Object}
\usage{
PawsGetSpectrumBatchRequest(antenna = NULL, capabilities = NULL,
  deviceDesc = NULL, locations = NULL, masterDeviceDesc = NULL,
  owner = NULL, requestType = NULL, type = NULL, version = NULL)
}
\arguments{
\item{antenna}{Depending on device type and regulatory domain, antenna characteristics may be required}

\item{capabilities}{The master device may include its device capabilities to limit the available-spectrum batch response to the spectrum that is compatible with its capabilities}

\item{deviceDesc}{When the available spectrum request is made on behalf of a specific device (a master or slave device), device descriptor information for the device on whose behalf the request is made is required (in such cases, the requestType parameter must be empty)}

\item{locations}{A geolocation list is required}

\item{masterDeviceDesc}{When an available spectrum batch request is made by the master device (a device with geolocation capability) on behalf of a slave device (a device without geolocation capability), the rules of the applicable regulatory domain may require the master device to provide its own device descriptor information (in addition to device descriptor information for the slave device in a separate parameter)}

\item{owner}{Depending on device type and regulatory domain, device owner information may be included in an available spectrum batch request}

\item{requestType}{The request type parameter is an optional parameter that can be used to modify an available spectrum batch request, but its use depends on applicable regulatory rules}

\item{type}{The message type (e}

\item{version}{The PAWS version}
}
\value{
PawsGetSpectrumBatchRequest object
}
\description{
PawsGetSpectrumBatchRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for a batch available spectrum query protocol.
}
\seealso{
Other PawsGetSpectrumBatchRequest functions: \code{\link{paws.getSpectrumBatch}}
}
