% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{PawsGetSpectrumBatchResponse}
\alias{PawsGetSpectrumBatchResponse}
\title{PawsGetSpectrumBatchResponse Object}
\usage{
PawsGetSpectrumBatchResponse(databaseChange = NULL, deviceDesc = NULL,
  geoSpectrumSchedules = NULL, maxContiguousBwHz = NULL,
  maxTotalBwHz = NULL, needsSpectrumReport = NULL, rulesetInfo = NULL,
  timestamp = NULL, type = NULL, version = NULL)
}
\arguments{
\item{databaseChange}{A database may include the databaseChange parameter to notify a device of a change to its database URI, providing one or more alternate database URIs}

\item{deviceDesc}{The database must return in its available spectrum response the device descriptor information it received in the master device's available spectrum batch request}

\item{geoSpectrumSchedules}{The available spectrum batch response must contain a geo-spectrum schedule list, The list may be empty if spectrum is not available}

\item{maxContiguousBwHz}{The database may return a constraint on the allowed maximum contiguous bandwidth (in Hertz)}

\item{maxTotalBwHz}{The database may return a constraint on the allowed maximum total bandwidth (in Hertz), which does not need to be contiguous}

\item{needsSpectrumReport}{For regulatory domains that require a spectrum-usage report from devices, the database must return true for this parameter if the geo-spectrum schedules list is not empty; otherwise, the database should either return false or omit this parameter}

\item{rulesetInfo}{The database should return ruleset information, which identifies the applicable regulatory authority and ruleset for the available spectrum batch response}

\item{timestamp}{The database includes a timestamp of the form, YYYY-MM-DDThh:mm:ssZ (Internet timestamp format per RFC3339), in its available spectrum batch response}

\item{type}{The message type (e}

\item{version}{The PAWS version}
}
\value{
PawsGetSpectrumBatchResponse object
}
\description{
PawsGetSpectrumBatchResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response message for the batch available spectrum query contains a schedule of available spectrum for the device at multiple locations.
}
