% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{PawsGetSpectrumRequest}
\alias{PawsGetSpectrumRequest}
\title{PawsGetSpectrumRequest Object}
\usage{
PawsGetSpectrumRequest(antenna = NULL, capabilities = NULL,
  deviceDesc = NULL, location = NULL, masterDeviceDesc = NULL,
  owner = NULL, requestType = NULL, type = NULL, version = NULL)
}
\arguments{
\item{antenna}{Depending on device type and regulatory domain, the characteristics of the antenna may be required}

\item{capabilities}{The master device may include its device capabilities to limit the available-spectrum response to the spectrum that is compatible with its capabilities}

\item{deviceDesc}{When the available spectrum request is made on behalf of a specific device (a master or slave device), device descriptor information for that device is required (in such cases, the requestType parameter must be empty)}

\item{location}{The geolocation of the master device (a device with geolocation capability that makes an available spectrum request) is required whether the master device is making the request on its own behalf or on behalf of a slave device (one without geolocation capability)}

\item{masterDeviceDesc}{When an available spectrum request is made by the master device (a device with geolocation capability) on behalf of a slave device (a device without geolocation capability), the rules of the applicable regulatory domain may require the master device to provide its own device descriptor information (in addition to device descriptor information for the slave device, which is provided in a separate parameter)}

\item{owner}{Depending on device type and regulatory domain, device owner information may be included in an available spectrum request}

\item{requestType}{The request type parameter is an optional parameter that can be used to modify an available spectrum request, but its use depends on applicable regulatory rules}

\item{type}{The message type (e}

\item{version}{The PAWS version}
}
\value{
PawsGetSpectrumRequest object
}
\description{
PawsGetSpectrumRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request message for the available spectrum query protocol which must include the device's geolocation.
}
\seealso{
Other PawsGetSpectrumRequest functions: \code{\link{paws.getSpectrum}}
}
