% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{PawsInitRequest}
\alias{PawsInitRequest}
\title{PawsInitRequest Object}
\usage{
PawsInitRequest(deviceDesc = NULL, location = NULL, type = NULL,
  version = NULL)
}
\arguments{
\item{deviceDesc}{The DeviceDescriptor parameter is required}

\item{location}{A device's geolocation is required}

\item{type}{The message type (e}

\item{version}{The PAWS version}
}
\value{
PawsInitRequest object
}
\description{
PawsInitRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The initialization request message allows the master device to initiate exchange of capabilities with the database.
}
\seealso{
Other PawsInitRequest functions: \code{\link{paws.init}}
}
