% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{PawsInitResponse}
\alias{PawsInitResponse}
\title{PawsInitResponse Object}
\usage{
PawsInitResponse(databaseChange = NULL, rulesetInfo = NULL, type = NULL,
  version = NULL)
}
\arguments{
\item{databaseChange}{A database may include the databaseChange parameter to notify a device of a change to its database URI, providing one or more alternate database URIs}

\item{rulesetInfo}{The rulesetInfo parameter must be included in the response}

\item{type}{The message type (e}

\item{version}{The PAWS version}
}
\value{
PawsInitResponse object
}
\description{
PawsInitResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The initialization response message communicates database parameters to the requesting device.
}

