% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{PawsNotifySpectrumUseRequest}
\alias{PawsNotifySpectrumUseRequest}
\title{PawsNotifySpectrumUseRequest Object}
\usage{
PawsNotifySpectrumUseRequest(deviceDesc = NULL, location = NULL,
  spectra = NULL, type = NULL, version = NULL)
}
\arguments{
\item{deviceDesc}{Device descriptor information is required in the spectrum-use notification message}

\item{location}{The geolocation of the master device (the device that is sending the spectrum-use notification) to the database is required in the spectrum-use notification message}

\item{spectra}{A spectrum list is required in the spectrum-use notification}

\item{type}{The message type (e}

\item{version}{The PAWS version}
}
\value{
PawsNotifySpectrumUseRequest object
}
\description{
PawsNotifySpectrumUseRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The spectrum-use notification message which must contain the geolocation of the Device and parameters required by the regulatory domain.
}
\seealso{
Other PawsNotifySpectrumUseRequest functions: \code{\link{paws.notifySpectrumUse}}
}
