% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{PawsNotifySpectrumUseResponse}
\alias{PawsNotifySpectrumUseResponse}
\title{PawsNotifySpectrumUseResponse Object}
\usage{
PawsNotifySpectrumUseResponse(type = NULL, version = NULL)
}
\arguments{
\item{type}{The message type (e}

\item{version}{The PAWS version}
}
\value{
PawsNotifySpectrumUseResponse object
}
\description{
PawsNotifySpectrumUseResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An empty response to the notification.
}
