% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{PawsRegisterRequest}
\alias{PawsRegisterRequest}
\title{PawsRegisterRequest Object}
\usage{
PawsRegisterRequest(antenna = NULL, deviceDesc = NULL, deviceOwner = NULL,
  location = NULL, type = NULL, version = NULL)
}
\arguments{
\item{antenna}{Antenna characteristics, including its height and height type}

\item{deviceDesc}{A DeviceDescriptor is required}

\item{deviceOwner}{Device owner information is required}

\item{location}{A device's geolocation is required}

\item{type}{The message type (e}

\item{version}{The PAWS version}
}
\value{
PawsRegisterRequest object
}
\description{
PawsRegisterRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The registration request message contains the required registration parameters.
}
\seealso{
Other PawsRegisterRequest functions: \code{\link{paws.register}}
}
