% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{PawsRegisterResponse}
\alias{PawsRegisterResponse}
\title{PawsRegisterResponse Object}
\usage{
PawsRegisterResponse(databaseChange = NULL, type = NULL, version = NULL)
}
\arguments{
\item{databaseChange}{A database may include the databaseChange parameter to notify a device of a change to its database URI, providing one or more alternate database URIs}

\item{type}{The message type (e}

\item{version}{The PAWS version}
}
\value{
PawsRegisterResponse object
}
\description{
PawsRegisterResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The registration response message simply acknowledges receipt of the request and is otherwise empty.
}

