% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{PawsVerifyDeviceRequest}
\alias{PawsVerifyDeviceRequest}
\title{PawsVerifyDeviceRequest Object}
\usage{
PawsVerifyDeviceRequest(deviceDescs = NULL, type = NULL, version = NULL)
}
\arguments{
\item{deviceDescs}{A list of device descriptors, which specifies the slave devices to be validated, is required}

\item{type}{The message type (e}

\item{version}{The PAWS version}
}
\value{
PawsVerifyDeviceRequest object
}
\description{
PawsVerifyDeviceRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The device validation request message.
}
\seealso{
Other PawsVerifyDeviceRequest functions: \code{\link{paws.verifyDevice}}
}
