% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{PawsVerifyDeviceResponse}
\alias{PawsVerifyDeviceResponse}
\title{PawsVerifyDeviceResponse Object}
\usage{
PawsVerifyDeviceResponse(databaseChange = NULL, deviceValidities = NULL,
  type = NULL, version = NULL)
}
\arguments{
\item{databaseChange}{A database may include the databaseChange parameter to notify a device of a change to its database URI, providing one or more alternate database URIs}

\item{deviceValidities}{A device validities list is required in the device validation response to report whether each slave device listed in a previous device validation request is valid}

\item{type}{The message type (e}

\item{version}{The PAWS version}
}
\value{
PawsVerifyDeviceResponse object
}
\description{
PawsVerifyDeviceResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The device validation response message.
}
