% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{RulesetInfo}
\alias{RulesetInfo}
\title{RulesetInfo Object}
\usage{
RulesetInfo(authority = NULL, maxLocationChange = NULL,
  maxPollingSecs = NULL, rulesetIds = NULL)
}
\arguments{
\item{authority}{The regulatory domain to which the ruleset belongs is required}

\item{maxLocationChange}{The maximum location change in meters is required in the initialization response, but optional otherwise}

\item{maxPollingSecs}{The maximum duration, in seconds, between requests for available spectrum}

\item{rulesetIds}{The identifiers of the rulesets supported for the device's location}
}
\value{
RulesetInfo object
}
\description{
RulesetInfo Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
This contains parameters for the ruleset of a regulatory domain that is communicated using the initialization and available-spectrum processes.
}
