% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_functions.R
\docType{package}
\name{spectrum_googleAuthR}
\alias{spectrum_googleAuthR}
\alias{spectrum_googleAuthR-package}
\title{Google Spectrum Database API
API for spectrum-management functions.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 23:47:28
filename: /Users/mark/dev/R/autoGoogleAPI/googlespectrumv1explorer.auto/R/spectrum_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item 
}
}

