% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{SpectrumMessage}
\alias{SpectrumMessage}
\title{SpectrumMessage Object}
\usage{
SpectrumMessage(bandwidth = NULL, frequencyRanges = NULL)
}
\arguments{
\item{bandwidth}{The bandwidth (in Hertz) for which permissible power levels are specified}

\item{frequencyRanges}{The list of frequency ranges and permissible power levels}
}
\value{
SpectrumMessage object
}
\description{
SpectrumMessage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Available spectrum can be logically characterized by a list of frequency ranges and permissible power levels for each range.
}
