% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{SpectrumSchedule}
\alias{SpectrumSchedule}
\title{SpectrumSchedule Object}
\usage{
SpectrumSchedule(eventTime = NULL, spectra = NULL)
}
\arguments{
\item{eventTime}{The event time expresses when the spectrum profile is valid}

\item{spectra}{A list of spectrum messages representing the usable profile}
}
\value{
SpectrumSchedule object
}
\description{
SpectrumSchedule Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The spectrum schedule element combines an event time with spectrum profile to define a time period in which the profile is valid.
}
