% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{Vcard}
\alias{Vcard}
\title{Vcard Object}
\usage{
Vcard(adr = NULL, email = NULL, fn = NULL, org = NULL, tel = NULL)
}
\arguments{
\item{adr}{The street address of the entity}

\item{email}{An email address that can be used to reach the contact}

\item{fn}{The full name of the contact person}

\item{org}{The organization associated with the registering entity}

\item{tel}{A telephone number that can be used to call the contact}
}
\value{
Vcard object
}
\description{
Vcard Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A vCard-in-JSON message that contains only the fields needed for PAWS:  - fn: Full name of an individual - org: Name of the organization - adr: Address fields - tel: Telephone numbers - email: Email addresses
}

