% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{VcardAddress}
\alias{VcardAddress}
\title{VcardAddress Object}
\usage{
VcardAddress(code = NULL, country = NULL, locality = NULL, pobox = NULL,
  region = NULL, street = NULL)
}
\arguments{
\item{code}{The postal code associated with the address}

\item{country}{The country name}

\item{locality}{The city or local equivalent portion of the address}

\item{pobox}{An optional post office box number}

\item{region}{The state or local equivalent portion of the address}

\item{street}{The street number and name}
}
\value{
VcardAddress object
}
\description{
VcardAddress Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The structure used to represent a street address.
}

