% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{VcardTelephone}
\alias{VcardTelephone}
\title{VcardTelephone Object}
\usage{
VcardTelephone(uri = NULL)
}
\arguments{
\item{uri}{A nested telephone URI of the form: tel:+1-123-456-7890}
}
\value{
VcardTelephone object
}
\description{
VcardTelephone Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The structure used to represent a telephone number.
}

