% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum_objects.R
\name{VcardTypedText}
\alias{VcardTypedText}
\title{VcardTypedText Object}
\usage{
VcardTypedText(text = NULL)
}
\arguments{
\item{text}{The text string associated with this item}
}
\value{
VcardTypedText object
}
\description{
VcardTypedText Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The structure used to represent an organization and an email address.
}
